requirements.txt
fastapi
pydantic
psycopg2-binary
databases
sqlalchemy
pyodbc
uvicorn
python-dotenv
alembic
passlib
pyjwt
python-jose[cryptography]
python-multipart
itsdangerous
fastapi-mail
pytest
httpx